requirements.txt
azure-identity
openai
python-dotenv
rich
dotenv-azd
aiohttp
pydantic-ai
httpx
PyGithub